backend/requirements_simple.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-generativeai>=0.8.3
qdrant-client>=1.12.1
python-dotenv>=1.0.1
pydantic>=2.9.2
pydantic-settings>=2.6.1
PyYAML>=6.0.2
cohere>=5.20.0
requests>=2.32.3
markdown>=3.7
SQLAlchemy>=2.0.35
PyJWT>=2.9.0
passlib>=1.7.4
bcrypt>=4.2.0
python-multipart>=0.0.12
redis>=5.2.0
google-auth>=2.40.1
google-auth-oauthlib>=1.2.3
googletrans>=3.0.0